+++ dashboard/src/components/SelectTeam.js
@@ -1,4 +1,3 @@
-/* eslint-disable react-hooks/exhaustive-deps */
 import React from 'react';
 import SelectCustom from './SelectCustom';